Receiver: fix faulty interpretation of flags sent by Sender
Sign-extension is performed when converting a byte to a char. This
will affect bits from 9 and up if its value is negative. So this will
corrupt flags if receiving 2 bytes (TransmitFlags.EXTENDED_FLAGS) and
if the value happens to be negative.

Fix by masking with 0xFF when converting the byte to a char as usual.

Also add debug logging of the value of flags in both Receiver and
Sender.