Add support for preserve ownership (-o, --owner)
If requested, this will try to update the ownership regardless of the
capabilities of the process. This does _not_ conform to the semantics
of rsync, which will only update ownership if the effective user id of
the process is 0.

NOTE: a side effect of chown in Linux is that setuid/setgid bits might
be cleared. yajsync does not try to recover from this, but rsync does.

NOTE: this will only change the ownership using the user name string.

TODO: fall back to update the user id if name change fails due to a
missing /etc/passwd entry.