Add support for legacy rsync safe file lists
Yay! Yajsync is now able to talk to rsync at kernel.org. More
precisely it talks to rsync 3.0.6, which is the default in rhel6 and
also with rsync 3.1.1pre2 which is the default in Ubuntu 14.04 (but
does not use safe file lists due to a bug (which is fixed in later
versions).