Fix issue #23: receiver: --owner is unable to handle incoming files w/ only uid and no name
Fixes https://github.com/perlundq/yajsync/issues/23

Receiver was not prepared to handle incoming ownership info with uid
but no user name. Fix by allowing User to contain the empty string as
the user name. If a target file lacks a user name and ownership needs
to be updated, update ownership by the uid.

NOTE: we do not try to detect whether uid updates are supported on the
system (if not an IOException will be thrown).