Fix issue #24: recursive module file listing of non dot directory lists the directory twice
Fixes https://github.com/perlundq/yajsync/issues/24

In recursive file listings the common case is to list the owning
directory for the segment before its contents, while excluding any
directories in the segment. The directories will be printed later with
their segment. But the initial segment does not have a directory; if
it is a dot dir, the first file is the dot dir containing the
remaining files. If it is not a dot dir, the next segment will be one
with the dir as owner. So the initial segment wrongly listed any
initial directory regardless if it is a dot dir or not.

Fixed by only listing the initial directory if it is a dot dir.