Fix issue #28: IAE when receiving existing file larger than 2 GiB
Fixes https://github.com/perlundq/yajsync/issues/28

When receiving matching content for an existing file larger than
2**31, an IllegalArgumentException occurs in FileChannel.read, due to
the offset being negative. The fileOffset is negative because of the
expression that computes a long fileOffset overflows before being
promoted to a long int. Fix by adding explicit cast to long of one
operand when computing the fileOffset.

TODO: add unit test for this