Sender: keep memory usage constant when recursing
Always keep the amount of active file list segments below a constant
amount. This will make sure Sender memory usage is roughly constant.

Note: we would rather work with files as units than with segments in
this case, but we cannot do this without risking deadlocks when
communicating with rsync.