Do not throw RuntimeInterruptException when iterating over file/module listings
Although the blocking iterator approach is easy to use it forces us to
throw an unchecked Exception (RuntimeInterruptException) when being
interrupted which is unorthodox and unexpected for the user.

Instead we provide a method take() in RsyncClient module/file
listings. This method retrieves the next available result and throws
InterruptedException.